Add team member component
